+++ blocks/footer/footer.js
@@ -1,5 +1,5 @@
 import { decorateIcons } from '../../scripts/lib-franklin.js';
-import { resolveRelativeURLs } from '../../scripts/scripts.js';
+import { resolveRelativeURLs } from '../../scripts/worker-commons.js';
 
 /**
 * loads and decorates the footer
@@ -8,15 +8,21 @@ import { resolveRelativeURLs } from '../../scripts/scripts.js';
 export default async function decorate(block) {
   // fetch footer content
   const resp = await fetch('https://www.walgreens.com/common/v1/footerui');
-  if (resp.ok) {
-    const data = await resp.json();
-    // decorate footer DOM
+  if (!resp.ok) {
+    return;
+  }
+
+  const worker = new Worker('../../scripts/absolute-worker.js');
+  const data = await resp.json();
+  // decorate footer DOM
+  worker.onmessage = (e) => {
     const footer = document.createElement('div');
-    footer.innerHTML = resolveRelativeURLs(data.content);
+    footer.innerHTML = e.data.content;
     decorateIcons(footer);
     block.append(footer);
     const footerStyles = document.createElement('style');
     footerStyles.innerHTML = data.clientLSGCSSContent;
     document.head.appendChild(footerStyles);
   }
+  worker.postMessage({ source: 'header', content: data.content });
 }
