+++ scripts/scripts.js
@@ -27,24 +27,6 @@ export function walgreensUrl(path) {
 }
 
 /**
- * Remaps the relative urls to absolute urls.
- * @param {string} content string of html with relative urls
- * @returns the string with absolute urls
- */
-export function resolveRelativeURLs(content) {
-  const baseUrl = 'https://walgreens.com';
-
-  // Use a regular expression to find relative links (starting with "/")
-  const relativeLinkRegex = /(?:href|action)="(?!\/images\/)(\/[^"]+)"/g;
-  const absoluteContent = content.replace(relativeLinkRegex, (match, relativePath) => {
-    // Combine the base URL and the relative path to create an absolute URL
-    const absoluteUrl = `${baseUrl}${relativePath}`;
-    return `href="${absoluteUrl}"`;
-  });
-  return absoluteContent;
-}
-
-/**
  * Adds the js and css to the head.
  * @param {JSON} fileList json object that comes with the UI API response
  */
@@ -52,7 +34,7 @@ export function loadFileList(fileList) {
   const baseUrl = 'https://www.walgreens.com';
 
   const skip = ['dtm', 'googleApi', 'speedIndex', 'lsgScriptMin'];
-  const eager = [ 'jquery', 'sly', 'headerSupport', 'lsgURL' ];
+  const eager = ['jquery', 'sly', 'headerSupport', 'lsgURL'];
 
   const scriptTags = document.querySelectorAll('script[src]');
